+++ examples/zapInDlmmIndirectSingleSided.ts
@@ -31,7 +31,10 @@ async function main() {
 
   const amountUseToAddLiquidity = new BN(0.001 * LAMPORTS_PER_SOL);
 
-  const zap = new Zap(connection, JUPITER_API_URL, JUPITER_API_KEY);
+  const zap = new Zap(connection, {
+    jupiterApiUrl: JUPITER_API_URL,
+    jupiterApiKey: JUPITER_API_KEY,
+  });
   const dlmm = await DLMM.create(connection, dlmmPool);
   const binDelta = 34;
   const singleSided = DlmmSingleSided.X; // MET
@@ -52,8 +55,10 @@ async function main() {
     maxDeltaId,
     strategy: StrategyType.Spot,
     singleSided,
-    jupiterApiUrl: JUPITER_API_URL,
-    jupiterApiKey: JUPITER_API_KEY,
+    config: {
+      jupiterApiUrl: JUPITER_API_URL,
+      jupiterApiKey: JUPITER_API_KEY,
+    },
   });
 
   const result = await zap.getZapInDlmmIndirectParams({